Bind mount shared /home in all containers
A directory on host is created under /var/tmp before starting the
containers. It is created using a systemd transient unit so it is owned
by root:.

The bind mount is setup by the firehpc systemd-nspawn wrapper with an
additional argument. The -U argument is removed as it enables
--private-users (ie. user namespace) which makes the binded mount point
owned by nobody not accessible by anyone (including root) in the
container.

The directory on host is removed in clean script using a transient unit
as well.