Generate random slurm mariadb password at runtime
There is now a generic bootstrap playbook which generate all sensitive
files (passwords and keys) in host local directory and setup SSH
service in all containers.

This playbook now calls the slurm tasks file bootstrap to generate the
munge key and a file containing the password of the Slurm MariaDB user.