feat(core): support SSH exec with library
Add support for running command on containers with SSH using paramiko
library, in addition to ssh executable binary. The library approach is
convenient when SSHClient is consumed by other programs, as it allows
better error handling and more control in commands output.

Paramiko is not great for handling interactive SSH sessions, TTYs must
be managed with additional and complex code. For this reason, the ssh
executable binary approach is kept for humans use-cases.