fix(core): optionally wait for IP addresses
Add optional wait boolean argument to Container.addresses() method to
wait for IP addresses to be available on a container. This is required
when IP addresses are retrieved right after the containers are started.
On the first try, the IP addresses are not yet assigned and the method
would return an empty list, which result in failures when applying
ansible configuration due to empty /etc/hosts files eventually.