refactor(conf): slurm nodes/partitions w/ RacksDB
Slurm nodes and partitions settings are now generated with the compute
nodes available in RacksDB database. FireHPC forwards the nodes to
Ansible with a new extra variables fhpc_nodes. The node list is
transformed into a string nodeset with a new Jinja2 filter nodeset_fold
that calls clustershell Nodeset class.