refactor!: replace zone→cluster
Replace the notion of zone in favor of cluster. This concerbs the CLI
options, configuration variables names, internal variables, and so on.

The notion of zone is specific to the virtual network management
features of systemd-nspawn. While FireHPC relies on this systemd-nspawn
feature to separate clusters' networks, it feels more natural to talk
about clusters at FireHPC level.