fix(conf): do not cgroups with Slurm in emulator
When used in frontend mode with FireHPC emulator mode, Slurm launches
all jobs on frontend nodes. If cgroups are enabled (the default), this
creates too many cgroups with files and dirs in the dedicated virtual
file system, up to a point where it is full and the system becomes
unstable.